Implemented step-by-step directions
Added a checkbox in the menu to toggle vibration and metric/imperial
units.

Updated the path-finding algorithm to connect a single line through
multiple collinear points in a given path.

Added a step-by-step direction system with distances in meters/feet and
direction.

Changed the UI of buttons to have icons with a transparent background.

Changed the way the current floor is displayed by updating the titlebar
text instead, moved the source/destination text to the top.